chore(cfg): refactor serde - see TODO
Config now can load from env vars + TOML config. Use the prefix
`AZ__` and the config key path with `.` replaced with `__` to
deserialise as that value.

e.g.
If we wanna increase the rates of epic drops, the key path in the toml
`Rate.DropItem.Epic` will be referenced as `AZ__Rate__DropItem__Epic`

BREAKING CHANGE: This commit is still a WIP for refactoring config
mgmt, ConfigMgr has a bunch of functions (related to key retrieval)
removed as we move towards favouring serde'ing the whole config
instead. This breaks authserver startup (in a compile error)

TODO:
1. remove default config files from `env/etc` + `env/etc/dist`
2. Add default toml config file generation based on serde
3. fix compile errors for authserver